requirements.txt
numpy>=1.23
scipy>=1.10
librosa>=0.10.1
soundfile>=0.12
scikit-learn>=1.3
hmmlearn>=0.3.0
umap-learn>=0.5.5
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.66
numba>=0.58









